split assertion logic test to follow the project conventions
